Move `DUPProperty` to `packets`
`DUPProperty` can be shared between MQTT-SN v.1 and v.2 packet
structures -> move it into common `packets` package.
